# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Blog",
        "description": "Listado de Endpoints de API Blog",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://api.codersfree.test"
        }
    ],
    "paths": {
...
        "/v1/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Listar todas los Roles",
                "operationId": "86eef45ce1339b1f7e2d84061168893e",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Admin"
                                                    },
                                                    "guard_name": {
                                                        "type": "string",
                                                        "example": "web"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Registrar un Rol",
                "operationId": "6d0fd19b2fcb9946ccd0b5adc50215a5",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "permissions": {
                                        "type": "array",
                                        "items": {}
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            ""
                        ]
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "access_token": {
                "type": "apiKey",
                "description": "Ingrese el token en formato(Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://api.codersfree.test/oauth/authorize",
                        "tokenUrl": "http://api.codersfree.test/oauth/token",
                        "refreshUrl": "http://api.codersfree.test/token/refresh",
                        "scopes": {
                            "create-post": "Crear un nuevo post",
                            "read-post": "Leer un post",
                            "update-post": "Actualizar un post",
                            "delete-post": "Eliminar un post",
                            "create-category": "Crear una nuevo categoria",
                            "read-category": "Leer una categoria",
                            "update-category": "Actualizar una categoria",
                            "delete-category": "Eliminar una categoria",
                            "create-tag": "Crear una nuevo etiqueta",
                            "read-tag": "Leer una etiqueta",
                            "update-tag": "Actualizar una etiqueta",
                            "delete-tag": "Eliminar una etiqueta"
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "passport": [
                "create-post",
                "read-post",
                "update-post",
                "delete-post",
                "create-category",
                "read-category",
                "update-category",
                "delete-category",
                "create-tag",
                "read-tag",
                "update-tag",
                "delete-tag"
            ]
        }
    ]
}